Moved Simple Feature Swtich tests to own class
