+++ Femah.Core.Tests/Femah.Core.Tests.csproj
@@ -63,6 +63,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PutApiRequestFactory.cs" />
     <Compile Include="RoleBasedFeatureSwitchTests.cs" />
+    <Compile Include="SimpleFeatureSwitchTests.cs" />
     <Compile Include="TestableFemahApiHttpHandler.cs" />
     <Compile Include="ProcessApiRequestTests.cs" />
   </ItemGroup>
